+++ packages/visual-editor/src/components/DropZone/Dropzone.tsx
@@ -1,6 +1,6 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { ElementType, useEffect, useMemo } from 'react';
 import { Droppable } from '@hello-pangea/dnd';
-import { ResolveDesignValueType } from '@/hooks/useBreakpoints';
+import type { ResolveDesignValueType } from '@contentful/experience-builder-core/types';
 import EditorBlock from './EditorBlock';
 import { ComponentData } from '@/types/Config';
 import { useTreeStore } from '@/store/tree';
@@ -13,15 +13,15 @@ import { EmptyEditorContainer } from '@components/EmptyContainer/EmptyContainer'
 import { getZoneParents } from '@/utils/zone';
 import { useZoneStore } from '@/store/zone';
 import { useDropZoneDirection } from '@/hooks/useDropZoneDirection';
+import { DESIGN_COMPONENT_NODE_TYPES } from '@contentful/experience-builder-core/constants';
 
 type DropZoneProps = {
   zoneId: string;
   node?: ComponentData;
   resolveDesignValue?: ResolveDesignValueType;
-  areEntitiesFetched?: boolean;
   className?: string;
   sectionId: string;
-  WrapperComponent?: any;
+  WrapperComponent?: ElementType | string;
 };
 
 function isDropEnabled(
@@ -31,8 +31,13 @@ function isDropEnabled(
   hoveringOverSection: boolean,
   draggingRootZone: boolean,
   isRootZone: boolean,
-  draggingOverArea: boolean
+  draggingOverArea: boolean,
+  isDesignComponent: boolean
 ) {
+  if (isDesignComponent) {
+    return false;
+  }
+
   if (!userIsDragging) {
     return false;
   }
@@ -57,7 +62,6 @@ export function DropZone({
   zoneId,
   sectionId,
   resolveDesignValue,
-  areEntitiesFetched,
   className,
   WrapperComponent = 'div',
   ...rest
@@ -78,8 +82,6 @@ export function DropZone({
   const draggedDestinationId = draggedItem && draggedItem.destination?.droppableId;
   const draggingParentIds = getZoneParents(draggedSourceId || '');
 
-  const [userWillDrag, setUserWillDrag] = useState(false);
-
   const hoveringRootZone = hoveringSection ? hoveringSection === zoneId : isRootZone;
   const hoveringOverZone = hoveringZone === zoneId;
 
@@ -90,7 +92,7 @@ export function DropZone({
 
   useEffect(() => {
     addSectionWithZone(sectionId);
-  }, [sectionId]);
+  }, [sectionId, addSectionWithZone]);
 
   const draggingOverArea = useMemo(() => {
     if (!userIsDragging) {
@@ -99,6 +101,8 @@ export function DropZone({
 
     return draggingParentIds[0] === zoneId;
   }, [userIsDragging, draggingParentIds, zoneId]);
+
+  const isDesignComponent = DESIGN_COMPONENT_NODE_TYPES.includes(node?.type || '');
 
   const draggingRootZone = draggedSourceId === ROOT_ID;
 
@@ -115,10 +119,11 @@ export function DropZone({
     hoveringOverSection,
     draggingRootZone,
     isRootZone,
-    draggingOverArea
+    draggingOverArea,
+    isDesignComponent
   );
 
-  if (!resolveDesignValue || !areEntitiesFetched) {
+  if (!resolveDesignValue) {
     return null;
   }
 
@@ -136,9 +141,9 @@ export function DropZone({
                 [styles.isEmpty]: isEmptyCanvas,
                 [styles.isRoot]: isRootZone,
                 [styles.hoveringRoot]: userIsDragging && hoveringRootZone,
-                [styles.isDragging]: userIsDragging,
+                [styles.isDragging]: userIsDragging && !isDesignComponent,
                 [styles.isHovering]: hoveringOverZone && !userIsDragging,
-                [styles.isDestination]: isDestination,
+                [styles.isDestination]: isDestination && !isDesignComponent,
               },
               className
             )}
@@ -164,10 +169,8 @@ export function DropZone({
                     key={componentId}
                     userIsDragging={userIsDragging}
                     draggingNewComponent={draggingNewComponent}
-                    setUserWillDrag={setUserWillDrag}
                     node={item}
                     resolveDesignValue={resolveDesignValue}
-                    areEntitiesFetched={areEntitiesFetched}
                   />
                 );
               })